+++ crates/prelude/CHANGELOG.md
@@ -10,7 +10,7 @@
 
 ### Minor
 
-- Add `crypto::{ecdsa,ecdh}` for ECDSA and ECDH
+- Add `crypto::{ecdsa,ecdh,ed25519}` for ECDSA, ECDH, and Ed25519
 - Add `gpio::{Event,Listener,Handler}` for GPIO event support
 - Add `platform::running_side()` function for the platform running side
 - Add `platform::opposite_version()` function for the platform version of the non-running side